requirements.txt
crewai>=0.37.0
fastapi>=0.115.0
uvicorn>=0.30.6
pymongo>=4.9.0
mongomock>=4.1.2
pydantic-settings>=2.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
requests>=2.32.0
tabula-py>=2.9.0
python-multipart>=0.0.6
tqdm>=4.66.0
python-dotenv>=1.0.0
minio==7.2.7
# VLMOCR Dependencies
# -------------------

# Async OpenAI client for vLLM API communication
openai>=2.9.0

# Data validation and serialization
pydantic>=2.12.5

# Retry logic with exponential backoff
tenacity>=9.1.0

# PDF processing
pdfplumber>=0.11.0

